Splits on semicolon, using regex
This aims to execute the current query by splitting the editor text on
semicolons. Simple quoting with "'" is meant to be supported. For
anything more complex (set delimiter, or Oracle-style q() quoting), a
more elaborate state machine would be called for.

I'm not myself 100% convinced what's better, newlines or semicolons.
Newlines could be a simple way to support (dense) PL/SQL blocks. Even
though for such use cases it should be bearable to simply use C-e C-e,
as before.