Execute query under cursor (#484)
* Steg0/executecurrent (#482)

* Two new commands and single-schema mode for oracle

Conflicts:

	packages/core/interface/connection.ts

* Extracted Run Current Query

* Splits on semicolon, using regex

This aims to execute the current query by splitting the editor text on
semicolons. Simple quoting with "'" is meant to be supported. For
anything more complex (set delimiter, or Oracle-style q() quoting), a
more elaborate state machine would be called for.

I'm not myself 100% convinced what's better, newlines or semicolons.
Newlines could be a simple way to support (dense) PL/SQL blocks. Even
though for such use cases it should be bearable to simply use C-e C-e,
as before.

* Improved regex

- capturing group unneeded, we use the whole match anyway
- mininum one char now -- simple way to avoid sending empty query

* New RE, to grab comments as well

* removed remnant

* use query utils to split SQL

this should be more in sync with existing C-e C-e behavior.
Simple implementation does not infer the dialect. But it should
already be better than the regex, quirks of the splitter
(which are independent) notwithstanding.

* Put import where most of the others are

Co-authored-by: Iosif Karaioannoglou <iosifak@gmail.com>
Co-authored-by: Matheus Teixeira <matheus.mtxr@gmail.com>

* cursor-query: highlight query under cursor

* cursor-query: fixed optional property on interface

Co-authored-by: Raimund Steger <rs@mytum.de>
Co-authored-by: Iosif Karaioannoglou <iosifak@gmail.com>