Fix Cassandra driver fail after creating UDF
There was some special logic in the CQL driver code that turned all objects
into strings. This would cause an error when an array (treated as an object
in Javascript) function was called; which would only happen when the
database had a user-defined function (UDF).

This commit removes this logic and adapts the Docker tests to include a
UDF.

Closes #495