Respect global telemetry.enableTelemetry setting
In the EU GDPR mandates that _all_ user data collection is opt-in.
This is currently not the case, and additionally sqltools does not
respect the global `telemetry.enableTelemetry` setting. The latter is
used by VS Code itself, and associated official extensions. Also some
bigger extensions, like GitLens respects that.

This fixes #90, again, regressed in:

https://github.com/mtxr/vscode-sqltools/commit/b13c8d7027246d908f7bcb32bccf5a5daff52e79