Add support to legacy CQL schemas
Upon connection to a CQL database, a query is run to check for compatibility
with a modern schema. If not found, compatibility with Cassandra versions prior
to 3.0 is assumed. This includes making the correct schema queries,
validator-to-type transformation, and not using `GROUP BY` to count columns
per table.

Aside from that, this commit includes improvements to driver code readability
and makes all columns explicit in queries.

Fixes #548